+++ lib/jekyll_sketchviz/generator.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+module JekyllSketchviz
+  class Generator
+    # TODO: Implement the generator logic.
+  end
+end