+++ lib/jekyll_sketchviz/graphviz_processor.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+module JekyllSketchviz
+  class GraphvizProcessor
+    # TODO: Implement Graphviz processing logic.
+  end
+end