+++ lib/jekyll_sketchviz/rough_transformer.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+module JekyllSketchviz
+  class RoughTransformer
+    # TODO: Add Rough.js transformations to SVG elements.
+  end
+end