+++ lib/jekyll_sketchviz/utilities/file_cache.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+module JekyllSketchviz
+  module Utilities
+    class FileCache
+      # TODO: Implement file caching utilities.
+    end
+  end
+end