+++ lib/jekyll_sketchviz/utilities/logger.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+module JekyllSketchviz
+  module Utilities
+    class Logger
+      # TODO: Implement logging utilities.
+    end
+  end
+end