+++ spec/rspec_setup_spec.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+require 'spec_helper'
+
+RSpec.describe RSpec do
+  it 'is properly set up' do
+    expect(1 + 1).to eq(2)
+  end
+end